Add Group APIs and resolveName
ResolveVanityURL is called resolveName in this API because the NAME is all that is needed, and 'vanity' is just extra Steam fluff that doesn't mean anything.

Group IDs are now useful with the custom getGroupInfo api, which is just the steam XML member list in disguise. It's stupid they have technically "deprecated" this but haven't replaced it with any other web API interface. Thanks Volvo.

Also minor bug fixes that I didn't notice through testing until now.